+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 * Initial release
+* Remove `format` property for string item
+* Add `pattern` property for string item